prueba base de datos con php

+++ PHP/eliminar_estudiante.php
@@ -1,6 +1,11 @@
 <?php
 include("conexion.php");
 
+if ($_SESSION['rol'] != 'admin') {
+    header("Location: listar_estudiantes.php");
+    exit;
+}
+
 $id = $_GET['id'];
 $conexion->query("DELETE FROM estudiantes WHERE id=$id");
